+++ frontend/src/components/Search/Search.js
@@ -1,11 +1,22 @@
-import React from 'react'
+import React, { useState } from 'react'
 import SearchBar from './SearchBar'
 
 function Search () {
-  
+
+  const [query, setQuery] = useState("")
+
+  const onChange = e => {
+    setQuery(e.target.value)
+  }
+
+  const onClick = e => {
+    e.preventDefault()
+    console.log(query)
+  }
+
   return (
     <>
-      <SearchBar />
+      <SearchBar onChange={onChange} onClick={onClick} />
     </>
 
   )
